#108 added frontend logic for multi Abrechnungs Groups added supporting files for the move
